Fixed : Verify token

+++ Api/controller/userContoller.js
@@ -1,6 +1,10 @@
 import { deleteUserService } from "../service/userService.js";
 
 export const deleteUser = async (req, res) => {
+  const user = await User.findById(req.params.id);
+  if (req.userId !== user._id.toString()) {
+    return res.status(403).send("You can only delete your account!");
+  }
   try {
     const result = await deleteUserService(req, res);
     if (result)
